# swagger/swagger.json
{"swagger":"2.0","info":{"title":"CryptoPages - Do business with fellow crypto enthusiasts.","description":"A directory of crypto enthusiasts offering their services. Written in goa.","contact":{"name":"Tom Utley","email":"tom@tomutley.com","url":"https://tomutley.com"},"version":""},"host":"localhost:8010","schemes":["http"],"consumes":["application/json"],"produces":["application/json"],"paths":{"/_ah/health":{"get":{"tags":["health"],"summary":"health health","description":"Perform health check.","operationId":"health#health","produces":["text/plain"],"responses":{"200":{"description":"OK"}},"schemes":["http"],"security":[{"jwt":[]}]}},"/js/{filepath}":{"get":{"summary":"Download public/js","operationId":"js#/js/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"],"security":[{"jwt":[]}]}},"/jwt/signin":{"post":{"tags":["jwt"],"summary":"signin jwt","description":"Creates a valid JWT","operationId":"jwt#signin","responses":{"204":{"description":"No Content","headers":{"Authorization":{"description":"Generated JWT","type":"string"}}},"401":{"description":"Unauthorized"}},"schemes":["http"],"security":[{"SigninBasicAuth":[]}]}},"/swagger.json":{"get":{"summary":"Download public/swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"],"security":[{"jwt":[]}]}},"/ui":{"get":{"summary":"Download public/html/index.html","operationId":"public#/ui","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"],"security":[{"jwt":[]}]}},"/user":{"get":{"tags":["user"],"summary":"search user","description":"Search users based on different criteria\n\nRequired security scopes:\n  * `api:access`","operationId":"user#search","produces":["application/vnd.cryptopages.user; type=collection","application/vnd.goa.error"],"parameters":[{"name":"coins","in":"query","description":"Search by crypto coins accepted (abbreviated)","required":false,"type":"string"},{"name":"jobCategory","in":"query","description":"Search by job Category","required":false,"type":"string","enum":["hardware","software","writing","legal","labor","automotive","services","others"]},{"name":"skills","in":"query","description":"Search by Skills","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CryptopagesUserCollection"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]},"post":{"tags":["user"],"summary":"create user","description":"Register new user","operationId":"user#create","produces":["application/vnd.goa.error"],"parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/CreateUserPayload"}}],"responses":{"201":{"description":"Resource created","headers":{"Location":{"description":"href to created resource","type":"string","pattern":"^/user/[0-9]+$"}}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/user/{username}":{"get":{"tags":["user"],"summary":"show user","description":"Get user by username\n\nRequired security scopes:\n  * `api:access`","operationId":"user#show","produces":["application/vnd.cryptopages.user","application/vnd.goa.error"],"parameters":[{"name":"username","in":"path","description":"username","required":true,"type":"string","minLength":1}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CryptopagesUser"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]},"delete":{"tags":["user"],"summary":"delete user","description":"Required security scopes:\n  * `api:access`","operationId":"user#delete","produces":["application/vnd.goa.error"],"parameters":[{"name":"username","in":"path","description":"username","required":true,"type":"string","minLength":2}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]},"patch":{"tags":["user"],"summary":"update user","description":"Required security scopes:\n  * `api:access`","operationId":"user#update","produces":["application/vnd.goa.error"],"parameters":[{"name":"username","in":"path","description":"The unique username for this person","required":true,"type":"string","minLength":2},{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/UserPayload"}}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]}}},"definitions":{"CreateUserPayload":{"title":"CreateUserPayload","type":"object","properties":{"available":{"type":"boolean","description":"Is this user available to provide work","example":true},"coins":{"type":"object","properties":{"bcc":{"type":"boolean","description":"Accepts Bitcoin Cash","example":false},"btc":{"type":"boolean","description":"Accepts Bitcoin","example":true},"eth":{"type":"boolean","description":"Accepts Ethereum","example":false},"ltc":{"type":"boolean","description":"Accepts Litecoin","example":false},"neo":{"type":"boolean","description":"Accepts Neo","example":true},"other":{"type":"object","properties":{"name":{"type":"string","description":"Name of the other coin accepted","example":"Perferendis totam voluptatibus facilis magni ut."}},"description":"Accepts some other coin","example":{"name":"Perferendis totam voluptatibus facilis magni ut."}},"xlm":{"type":"boolean","description":"Accepts Lumen","example":true},"xrp":{"type":"boolean","description":"Accepts Ripple","example":false}},"description":"The coins this user will accept as payment","example":{"bcc":false,"btc":true,"eth":false,"ltc":false,"neo":true,"other":{"name":"Perferendis totam voluptatibus facilis magni ut."},"xlm":true,"xrp":false}},"email":{"type":"object","properties":{"makePublic":{"type":"boolean","description":"Should the email address be shown to others","example":true},"value":{"type":"string","description":"the email address","example":"me@someplace.com","format":"email","minLength":5}},"description":"The user's email address","example":{"makePublic":true,"value":"me@someplace.com"}},"jobCategory":{"type":"string","example":"legal","enum":["hardware","software","writing","legal","labor","automotive","services","others"]},"jobDescription":{"type":"string","example":"I'm looking for small remote projects"},"location":{"type":"object","properties":{"makePublic":{"type":"boolean","description":"Should the location be shown to others","example":true},"value":{"type":"string","description":"the location","example":"Charleston, SC, USA","minLength":2}},"description":"the user's location geographically","example":{"makePublic":true,"value":"Charleston, SC, USA"}},"name":{"type":"string","description":"The user's full name","example":"John Smith","minLength":2},"password":{"type":"string","description":"A password (only exposed to user)","example":"somethingreallyhardtoguess123%$!@#","minLength":8},"skills":{"type":"array","items":{"type":"string","example":"Cupiditate praesentium."},"example":["Cupiditate praesentium."]},"username":{"type":"string","description":"The unique username for this person","example":"scubasteve","minLength":2}},"example":{"available":true,"coins":{"bcc":false,"btc":true,"eth":false,"ltc":false,"neo":true,"other":{"name":"Perferendis totam voluptatibus facilis magni ut."},"xlm":true,"xrp":false},"email":{"makePublic":true,"value":"me@someplace.com"},"jobCategory":"legal","jobDescription":"I'm looking for small remote projects","location":{"makePublic":true,"value":"Charleston, SC, USA"},"name":"John Smith","password":"somethingreallyhardtoguess123%$!@#","skills":["Cupiditate praesentium."],"username":"scubasteve"},"required":["name","username","password"]},"CryptopagesUser":{"title":"Mediatype identifier: application/vnd.cryptopages.user; view=default","type":"object","properties":{"available":{"type":"boolean","description":"Is this user available to provide work","example":true},"coins":{"type":"object","properties":{"bcc":{"type":"boolean","description":"Accepts Bitcoin Cash","example":false},"btc":{"type":"boolean","description":"Accepts Bitcoin","example":true},"eth":{"type":"boolean","description":"Accepts Ethereum","example":false},"ltc":{"type":"boolean","description":"Accepts Litecoin","example":false},"neo":{"type":"boolean","description":"Accepts Neo","example":true},"other":{"type":"object","properties":{"name":{"type":"string","description":"Name of the other coin accepted","example":"Perferendis totam voluptatibus facilis magni ut."}},"description":"Accepts some other coin","example":{"name":"Perferendis totam voluptatibus facilis magni ut."}},"xlm":{"type":"boolean","description":"Accepts Lumen","example":true},"xrp":{"type":"boolean","description":"Accepts Ripple","example":false}},"description":"The coins this user will accept as payment","example":{"bcc":false,"btc":true,"eth":false,"ltc":false,"neo":true,"other":{"name":"Perferendis totam voluptatibus facilis magni ut."},"xlm":true,"xrp":false}},"email":{"type":"object","properties":{"makePublic":{"type":"boolean","description":"Should the email address be shown to others","example":true},"value":{"type":"string","description":"the email address","example":"me@someplace.com","format":"email","minLength":5}},"description":"The user's email address","example":{"makePublic":true,"value":"me@someplace.com"}},"href":{"type":"string","description":"API href for making requests on the user","example":"/users/scubasteve"},"jobCategory":{"type":"string","example":"services","enum":["hardware","software","writing","legal","labor","automotive","services","others"]},"jobDescription":{"type":"string","example":"I'm looking for small remote projects"},"location":{"type":"object","properties":{"makePublic":{"type":"boolean","description":"Should the location be shown to others","example":true},"value":{"type":"string","description":"the location","example":"Charleston, SC, USA","minLength":2}},"description":"the user's location geographically","example":{"makePublic":true,"value":"Charleston, SC, USA"}},"name":{"type":"string","description":"The user's full name","example":"John Smith","minLength":2},"password":{"type":"string","description":"A password (only exposed to user)","example":"somethingreallyhardtoguess123%$!@#","minLength":8},"skills":{"type":"array","items":{"type":"string","example":"Cupiditate praesentium."},"example":["Cupiditate praesentium.","Cupiditate praesentium."]},"username":{"type":"string","description":"The unique username for this person","example":"scubasteve","minLength":2}},"description":"A cryptopages user (default view)","example":{"available":true,"coins":{"bcc":false,"btc":true,"eth":false,"ltc":false,"neo":true,"other":{"name":"Perferendis totam voluptatibus facilis magni ut."},"xlm":true,"xrp":false},"email":{"makePublic":true,"value":"me@someplace.com"},"href":"/users/scubasteve","jobCategory":"services","jobDescription":"I'm looking for small remote projects","location":{"makePublic":true,"value":"Charleston, SC, USA"},"name":"John Smith","password":"somethingreallyhardtoguess123%$!@#","skills":["Cupiditate praesentium.","Cupiditate praesentium."],"username":"scubasteve"},"required":["username","href","name","jobCategory"]},"CryptopagesUserCollection":{"title":"Mediatype identifier: application/vnd.cryptopages.user; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/CryptopagesUser"},"description":"CryptopagesUserCollection is the media type for an array of CryptopagesUser (default view)","example":[{"available":true,"coins":{"bcc":false,"btc":true,"eth":false,"ltc":false,"neo":true,"other":{"name":"Perferendis totam voluptatibus facilis magni ut."},"xlm":true,"xrp":false},"email":{"makePublic":true,"value":"me@someplace.com"},"href":"/users/scubasteve","jobCategory":"services","jobDescription":"I'm looking for small remote projects","location":{"makePublic":true,"value":"Charleston, SC, USA"},"name":"John Smith","password":"somethingreallyhardtoguess123%$!@#","skills":["Cupiditate praesentium.","Cupiditate praesentium."],"username":"scubasteve"},{"available":true,"coins":{"bcc":false,"btc":true,"eth":false,"ltc":false,"neo":true,"other":{"name":"Perferendis totam voluptatibus facilis magni ut."},"xlm":true,"xrp":false},"email":{"makePublic":true,"value":"me@someplace.com"},"href":"/users/scubasteve","jobCategory":"services","jobDescription":"I'm looking for small remote projects","location":{"makePublic":true,"value":"Charleston, SC, USA"},"name":"John Smith","password":"somethingreallyhardtoguess123%$!@#","skills":["Cupiditate praesentium.","Cupiditate praesentium."],"username":"scubasteve"}]},"UserPayload":{"title":"UserPayload","type":"object","properties":{"available":{"type":"boolean","description":"Is this user available to provide work","example":true},"coins":{"type":"object","properties":{"bcc":{"type":"boolean","description":"Accepts Bitcoin Cash","example":false},"btc":{"type":"boolean","description":"Accepts Bitcoin","example":true},"eth":{"type":"boolean","description":"Accepts Ethereum","example":false},"ltc":{"type":"boolean","description":"Accepts Litecoin","example":false},"neo":{"type":"boolean","description":"Accepts Neo","example":true},"other":{"type":"object","properties":{"name":{"type":"string","description":"Name of the other coin accepted","example":"Perferendis totam voluptatibus facilis magni ut."}},"description":"Accepts some other coin","example":{"name":"Perferendis totam voluptatibus facilis magni ut."}},"xlm":{"type":"boolean","description":"Accepts Lumen","example":true},"xrp":{"type":"boolean","description":"Accepts Ripple","example":false}},"description":"The coins this user will accept as payment","example":{"bcc":false,"btc":true,"eth":false,"ltc":false,"neo":true,"other":{"name":"Perferendis totam voluptatibus facilis magni ut."},"xlm":true,"xrp":false}},"email":{"type":"object","properties":{"makePublic":{"type":"boolean","description":"Should the email address be shown to others","example":true},"value":{"type":"string","description":"the email address","example":"me@someplace.com","format":"email","minLength":5}},"description":"The user's email address","example":{"makePublic":true,"value":"me@someplace.com"}},"jobCategory":{"type":"string","example":"legal","enum":["hardware","software","writing","legal","labor","automotive","services","others"]},"jobDescription":{"type":"string","example":"I'm looking for small remote projects"},"location":{"type":"object","properties":{"makePublic":{"type":"boolean","description":"Should the location be shown to others","example":true},"value":{"type":"string","description":"the location","example":"Charleston, SC, USA","minLength":2}},"description":"the user's location geographically","example":{"makePublic":true,"value":"Charleston, SC, USA"}},"name":{"type":"string","description":"The user's full name","example":"John Smith","minLength":2},"password":{"type":"string","description":"A password (only exposed to user)","example":"somethingreallyhardtoguess123%$!@#","minLength":8},"skills":{"type":"array","items":{"type":"string","example":"Cupiditate praesentium."},"example":["Cupiditate praesentium."]},"username":{"type":"string","description":"The unique username for this person","example":"scubasteve","minLength":2}},"example":{"available":true,"coins":{"bcc":false,"btc":true,"eth":false,"ltc":false,"neo":true,"other":{"name":"Perferendis totam voluptatibus facilis magni ut."},"xlm":true,"xrp":false},"email":{"makePublic":true,"value":"me@someplace.com"},"jobCategory":"legal","jobDescription":"I'm looking for small remote projects","location":{"makePublic":true,"value":"Charleston, SC, USA"},"name":"John Smith","password":"somethingreallyhardtoguess123%$!@#","skills":["Cupiditate praesentium."],"username":"scubasteve"}},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"NoContent":{"description":"No Content"},"NotFound":{"description":"Not Found"},"OK":{"description":"OK","schema":{"$ref":"#/definitions/CryptopagesUser"}},"Unauthorized":{"description":"Unauthorized"}},"securityDefinitions":{"SigninBasicAuth":{"type":"basic"},"jwt":{"type":"apiKey","description":"\n\n**Security Scopes**:\n  * `api:access`: API access","name":"Authorization","in":"header"}}}